local/requirements.txt
# ローカル版 vol.1 の依存パッケージ
# スキャンPDFもOCRで扱いたい場合は [hybrid] extra を残したままにしてください。
opendataloader-pdf[hybrid]
PyYAML>=6.0
